Bug #1409038 thermal: sprd: fix variable type overflow problem
[root cause  ] overflow problem
[changes     ] sprd_gpu_device.c
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums512
[test case]:power on/off
[reviewers   ] vincent.wang@unisoc.com,jeson.gao@unisoc.com
[change_type ] feature_bugfix
[tag_product ] common

Inappropriate variable type may cause overflow problem,so modified these types.

+++ drivers/thermal/sprd_gpu_device.c
@@ -142,7 +142,7 @@ static unsigned long get_static_power(struct devfreq *df, unsigned long m_volt)
 	struct thermal_cooling_device *tcd;
 	unsigned long t_scale, v_scale;
 	u32 gpu_coeff;
-	u32 tmp_power = 0;
+	unsigned long tmp_power = 0;
 	unsigned long power;
 	int gpu_cluster = GPU_CLUSTER_ID;
 	int gpu_cores = GPU_CORE_NUM;
